move the util package into the domain package.
