+++ src/test/kotlin/domain/usecase/auth/RegisterUseCaseTest.kt
@@ -9,7 +9,7 @@ import org.example.entity.User
 import org.example.entity.UserType
 import org.example.logic.repository.AuthenticationRepository
 import org.example.logic.usecase.auth.RegisterUseCase
-import org.example.utils.PlanMateException
+import domain.utils.exception.PlanMateException
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
