+++ src/test/kotlin/domain/usecase/project/AddProjectUseCaseTest.kt
@@ -10,7 +10,7 @@ import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.example.logic.repository.ProjectRepository
 import org.example.logic.usecase.project.AddProjectUseCase
-import org.example.utils.PlanMateException
+import domain.utils.exception.PlanMateException
 import org.junit.jupiter.api.assertThrows
 import kotlin.test.Test
