+++ src/test/kotlin/domain/usecase/state/GetStatesByProjectIdTest.kt
@@ -7,7 +7,7 @@ import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.example.logic.repository.WorkflowStateRepository
 import org.example.logic.usecase.state.GetStatesByProjectId
-import org.example.utils.PlanMateException
+import domain.utils.exception.PlanMateException
 import org.junit.jupiter.api.BeforeEach
 import java.util.*
 import kotlin.test.Test
